+++ SpecFlow.VisualStudio/ProjectSystem/NuGetVersion.cs
@@ -1,16 +1,28 @@
 ﻿using System;
+using System.Text.RegularExpressions;
+using SpecFlow.VisualStudio.Annotations;
 
 namespace SpecFlow.VisualStudio.ProjectSystem
 {
-    public class NuGetVersion
+    public record NuGetVersion
     {
         public Version Version { get; }
         public string PreReleaseSuffix { get; }
         public bool IsPrerelease => PreReleaseSuffix != null;
+        /// <summary>The project's requested package range for the package.</summary>
+        /// <remarks>
+        /// If the project uses packages.config, this will be same as the installed package version.
+        /// If the project uses PackageReference, this is the version string in the project file, which may not match the resolved package version, and may not be single version string.
+        /// If the project uses PackageReference, and the package is a transitive dependency, the value will be null.
+        /// </remarks>
+        public string RequestedRange { get; }
 
-        public NuGetVersion(string versionSpecifier)
+        public bool IsFloating => Regex.IsMatch(RequestedRange, @"\*");
+
+        public NuGetVersion(string versionSpecifier, [NotNull]string requestedRange)
         {
             if (versionSpecifier == null) throw new ArgumentNullException(nameof(versionSpecifier));
+            RequestedRange = requestedRange;
 
             var versionParts = versionSpecifier.Split(new[] {'-'}, 2);
             if (Version.TryParse(versionParts[0], out var version))
@@ -23,33 +35,14 @@ namespace SpecFlow.VisualStudio.ProjectSystem
 
         public override string ToString()
         {
-            return IsPrerelease ? $"{Version}-{PreReleaseSuffix}" : Version.ToString();
+            var str = IsPrerelease ? $"{Version}-{PreReleaseSuffix}" : Version.ToString();
+            if (IsFloating) str += $"({RequestedRange})";
+            return str;
         }
 
         public string ToShortVersionString()
         {
             return $"{Version.Major}{Version.Minor:00}{Version.Build}";
-        }
-
-        protected bool Equals(NuGetVersion other)
-        {
-            return Equals(Version, other.Version) && string.Equals(PreReleaseSuffix, other.PreReleaseSuffix);
-        }
-
-        public override bool Equals(object obj)
-        {
-            if (ReferenceEquals(null, obj)) return false;
-            if (ReferenceEquals(this, obj)) return true;
-            if (obj.GetType() != this.GetType()) return false;
-            return Equals((NuGetVersion) obj);
-        }
-
-        public override int GetHashCode()
-        {
-            unchecked
-            {
-                return ((Version != null ? Version.GetHashCode() : 0) * 397) ^ (PreReleaseSuffix != null ? PreReleaseSuffix.GetHashCode() : 0);
-            }
         }
     }
 }