+++ Changelog.md
@@ -1,5 +1,7 @@
 # v$vNext$ - $buildDate$
 
+* Fix: Support floating versions of SpecFlow package references
+
 # v2021.0.187 - 2021-09-24
 
 * Fix: Could not load file or assembly error when opening ASP.NET MVC test projects.